+++ CHANGELOG.md
@@ -9,4 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Initializing repo
+- `namespace` prop on the `ContentWrapper` component.
+
+### Added
+
+- Initializing repo.